refactor: extract useExternalLinkFlow hook (#3501 Phase 1i)
